(related? tdf#155177) sw: layout: fix another RemoveFollowFlowLine SNAFU
This happens with the following commit, on testForcepoint76, for no
obvious reason.

A follow table frame is in SwTabFrame::MakeAll(), then its follow flow
line becomes empty and it does SetRemoveFollowFlowLinePending() on the
master, then it formats its follow flow line (but at that time there's
again a text frame in it), then that content moves backward, formats its
uppers until it gets to the master table frame, which wants to join the
row frame that's being formatted on the stack.

Obviously a SwFrameDeleteGuard in lcl_RecalcRow() can't hurt, but it
needs an additional change later to prevent it from moving rows of the
follow back.

Change-Id: Idca90b7eb67994385ccd306562ea26ff39e884f9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156117
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
(cherry picked from commit 8c24bc5baa1f501943bf955e07787d80cb7a5763)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156146
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>