Change theme location to SdrModel and SdrPage (master page only)
In Writer and Calc we only have one theme per document, so the
theme can be stored inside the SdrModel. For Imporess/Draw the
theme is defined per master page, so it needs to be stored on the
master page. This changes the implementation to reflect that.

In addition refactor all the usages in writer and calc to get and
set the theme to SdrModel, when needed and rename the methods on
the SdrPageProperties from {Get,Set}Theme to {get,set}Theme.

Change-Id: I0cc3b332c029c28d15f8cda748c578c2a54a5c61
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156128
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit d150b7a60d7be4b09a838ae4c038e78509f75126)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156175
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Ashod Nakashian <ash@collabora.com>