remove null checks for mpObj in UIObjectUnoObj
Checking for mpObj == nullptr implies that in some sense this object
could end up in that state, but the only way that can happen is if we
call this object after its destructor has been called. In which case
throwing RuntimeException is not appropriate because calling a dead
object should not be possible.

Rather just let it SIGSEGV instead of hiding it behind an exception that
might be ignored.

Change-Id: I92911a3108f74ed5dfa4986bec028cbc365ac916
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156345
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>