tdf#104833 a11y: Use VCLXAccessibleEdit for VCLXMultiLineEdit
Instead of the generic `VCLXAccessibleComponent`,
use `VCLXAccessibleEdit` as a11y class for
`VCLXMultiLineEdit` (the component interface
class for MultiLineEdit) as well, not just for `VCLXEdit`.

With the preparatory changes from

    Change-Id I8218db61feb07605f6ea5309f26eebd38312458a
    tdf#104833 a11y: Don't use VCLXEdit in VCLXAccessibleEdit

in place, it behaves as expected when interacting
with the multi line edit from the sample dialog in
attachment 189287 in tdf#104833 using Accerciser and
the qt6 or gtk3 VCL plugin (and the text can also be
edited when removing the readonly flag from the control)
or the update dialog (which uses UNO controls,
`UnoControlEditModel` for the panes that were not
announced, s. extensions/source/update/check/updatehdl.cxx ).

In particular, the a11y text interface is supported by
the multi line edit on the a11y layer, which makes
Orca with the gtk3 VCL plugin announce the text content
when the control gets focused.

NVDA on Windows still announces "Checking..." instead
of using the actual text that would be retrievable via
the IAccessibleText interface now. This is probably
because "Checking..." is the (outdated) accessible name,
as can be seen by interacting with the object in NVDA's
Python console:

    >>> focus.name
    'Checking...'
    >>> txt = focus.IAccessibleTextObject
    >>> txt.text(0, txt.nCharacters)
    'LibreOfficeDev 24.2 is up to date.'

For both, gtk3 and qt6, the a11y object does have the
new text as a11y name as well, as can be verified in
Accerciser:

    In [10]: acc.name
    Out[10]: 'LibreOfficeDev 24.2 is up to date.'
    In [11]: txt = acc.queryText()
    In [12]: txt.getTextAtOffset(0,3)
    Out[12]: ('LibreOfficeDev 24.2 is up to date.', 0, 34)

Change-Id: Id80e191cdd5342b3215fdb9d2ad3847470366337
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156402
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>