tdf#147704 fix reading TIFF metadata (width and height)
When we prepare (not read and load) the bitmap we need to read
only the metadata - most importantly pixel width and height. This
didn't work correctly for the TIFF files or more importantly the
specific TIFF files in the document, and the result was that the
size reported was 0,0 and somehow the crop factor wasn't applied
correctly. This change fixes the metadata reading and makes it
more robust.

Change-Id: I60951e5d01afc311c400589496ae82377ccb4cf5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156457
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit 9c636d76a9a7e167da39913ab60f5135f8e831a6)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156437
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>