ScriptForge (SF_Document) new XDocumentSettings property
The XDocumentSettings property returns a
   com.sun.star.XXX.DocumentSettings
UNO object.

XXX = sheet, text, drawing or presentation

It gives access to a bunch of UNO internal
properties, specific to the document(s type.

The property is available in Basic and Python
user scripts.

An update of the SF_Document service help
page is required.

Change-Id: I9d6db473c91ac5b1814def9cd100e874aa5490cd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156475
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins