tdf#156318 sw floattable: fix follow text flow handling on interactive edit
Once the bugdoc has 1 col for pages, setting the top and bottom margins
to 2.5cm created a split fly, but the follow fly was on page 3, not on
page 2.

What happened is that the layout was already wrong before calling into
SwFrame::GetNextFlyLeaf() (that does the split), the "follow text flow"
feature shifted the fly down (top was originally 1341, then the
increased margin means 2475, but an additional 4329 was set), and then
the fly was created on the wrong page.

Fix the problem by not calling GetAnchoredObj().SetObjTop() in
SwToContentAnchoredObjectPosition::CalcPosition() when the fly doesn't
fit anymore, this way the fly can shrink, a follow can be created and
some of the content will be moved to the follow fly, leading to a
correct split fly on pages 1 & 2.

This also fixes the previously appearing "fly frame has negative height
now" warnings on setting the page margin.

Change-Id: I75bf9cfa19dc5c7ce9607417d6a4121b8e85ac47
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156506
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
(cherry picked from commit 401c175d1bbe3c64e5dd96e3b23779999271cfb1)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156531
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>