websocket URP flush hangs with large buffer
over 4k hits limit so flush waits until something reads, but the
read is in the same thread after the flush.

We are just using _outputStream to accumulate the bytes to be send
so change this to use a ByteArrayOutputStream and take over its bytes
and clear it when we flush.

Change-Id: I17b90e1c7d4302f153b5832e60ac0e0f2b86ace9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156565
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>