starmath: Keep focus on math widget when inline editing is enabled
Since originally all editing happened in the edit window, it was made to
grab focus instead of the math widget, but we don’t want that anymore
when inline editing is enable as we want to edit everything in the math
widget.

This also removes ancient test that is now failing. We will eventually
drop the edit window altogether.

Change-Id: Ia1628a454bbbdd923ff582564eca1ab4ec830f38
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156579
Tested-by: Jenkins
Reviewed-by: خالد حسني <khaled@libreoffice.org>