tdf#156959 - Remove pinned icons from thumbnail view item
Remove pinned icons from thumbnail view item and pushed it down
to the recent documents view item. This avoids that the pinned
icon will be shown in a simple thumbnail view, e.g., in the
thumbnail view of the side pane in base.

In addition, fixed an error with the number of elements in the
thumbnail view in the start center when items are pinned, i.e.,
removed a spurious new line.

Change-Id: I76026bfa9239f709aa60ff12efabe28cb09f0d76
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156625
Tested-by: Jenkins
Reviewed-by: Andreas Heinisch <andreas.heinisch@yahoo.de>