Improve the HasFoldedOutlineContentSelected function
Return false immediatly if there are no outline nodes in the document
since then there can be no folded outline content selected.

Adjust selection that includes the end of content node so that the
end of content node is not included in the selection. This fixes a
crash that occurs when outline folding is enabled and Master
document 'Text' is deleted by using the Navigator popup menu item
'Delete'.

Change-Id: I28be127566c3820b3e78834495bbb354cf2d01aa
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156631
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>