tdf#157137 a11y: Don't report active descendant when not focused
Since an active descendant is the UI element with keyboard focus,
only send an `ACTIVE_DESCENDANT_CHANGED` event in `VCLXAccessibleList`
if the listbox/combobox actually has focus, not when the selection
changed otherwise, e.g. because the value was changed programmatically
because it depends on another UI control (like the paper size
in the print dialog depends on the selected printer.

Change-Id: I008e742ccf3b81c70832707b2f3a6551827a0706
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156666
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>