tdf#154063 allow autocorrect GtkTreeView to shrink
It will allow shrinking if there is a horizontal scrollbar.

We don't actually want to show a horizontal scrollbar so use "external"
which vcl will ignore, and gtk not show a visible scrollbar but
allow scrolling. So when we shrink it will "scroll" and the resize
handler sets the column widths as we want and it works out to behave
like vcl does.

If this fails to work out for some reason we could put this inside
some other container which can shrink as an alternative solution.

Change-Id: Ic7e3ab4bff344f4ed5c1a6317644e697c9cfc395
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156673
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>