tdf#137595 sw: current fly might not be selected
Likely true with draw flies.

In the noticed cases, sometimes the fly would be selected
(and then things worked as expected),
but if it was not selected or if the fly was changed,
then it just became the current fly, but not selected.

So, check for both. Since this function already occurs,
likely this draw/writer synchronization has already
bitten a lot of cases. Here is one more...

Change-Id: I238027489e6e2cc49e7503c1f93779d1e96dbdcf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156786
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>