sw floattable, nesting: fix overlap support
The bugdoc had a nested, split floating table with overlap=never and
that lead to a layout loop.

The root of the trouble seems to be that the inner fly will obviously
overlap with its outer fly but we tried to prevent that and we failed.

Fix the problem by ignoring inner flys in
SwToContentAnchoredObjectPosition::CalcOverlap().

This also allows removing special handling of nested split flys in
lcl_ArrangeLowers() and special handling of split flys at an other place
in the same function, because now the non-bDirectMove case works out of
the box.

Change-Id: Icf3a8f776aa758ef4ae2c2994a7216c5a6142a62
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156810
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins