tdf#157173 Enlarging Fontwork shape replaces used font
regression from
    commit b153d147c9d813f78dd551fc186c261f52c929a1
    Author: Noel Grandin <noel.grandin@collabora.co.uk>
    Date:   Wed Apr 28 15:27:14 2021 +0200
    remove support for BITMASK in vcl backends

So the code that I removed had
    if (meFormat == DeviceFormat::BITMASK)
        SetAntialiasing( AntialiasingFlags::DisableText );
in the
    VirtualDevice::ImplInitVirDev
method.

Possibly we will need to make similar fixes in other places that used
BITMASK, but this looks like the only spot that really needs it.

No idea exactly how this line of code relates to the observed bug
though.

Change-Id: I858f15cf70a452568128d5ccccb365d9a11f7062
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156815
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>