sc: row import - default style index is 0 and not -1
When importing CT_Row structure, the style index is 0 by default
and not -1. The effect of this is that the style is not applied
for the rows, because the index is -1. With the change to 0, the
style applied is the first style, which is the correct behaviour
and this fixes the issue with the test document.

Change-Id: If5ab571e8dd7ec1319e7cb4d26ba006ca827eca6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156845
Tested-by: Tomaž Vajngerl <quikee@gmail.com>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit 0da88ee88255083ec9557ff25005a745dcdbef14)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156877
Tested-by: Jenkins
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>