Revert "loplugin:unodispose in binaryurp"
This reverts commit 3d9f9b1b299251c5c28d74e951b4447193ad7ade.

Reason for revert: comment from sberg:

I'm not convinced that a general rule of "clear all references in dispose()" is a good idea.  For example, in this class, access to connection_ and provider_ is not guarded against MT-races, on the basis that those two are never modified post construction.  That is broken now.  Also, e.g. Bridge::getDescription dereferences connection_, and benignly used to work fine even when called on a disposed instance.  Now, it causes UB.

Change-Id: I6039d5b1488146f09439312800ceb1de3cfcf0ba
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156878
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>