tdf#157223 get sequence name before toggling checkboxes
The expectation is that setting state/values on widgets from code
doesn't trigger any callbacks, which are assumed to be from user
interaction only.

But explictly toggling on a radio implicitly toggles off the other
radio, and while there is no callback from the explicit toggle on,
there is from the implicit toggle off.

Just move the explicit toggles to the end, after where the
values are read which the callback affects.

Change-Id: Icfc2c94950be671d6fc23789584b4d3c6e5395d8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156879
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
Tested-by: Jenkins