Make sure bin/run script finds the correct dlls
On Windows/Cygwin, if you put "C:\Program Files\LibreOffice\program"
in PATH, bin/run will have trouble finding correct vcllo.dll and other
LibreOffice .dll files inside instdir/program. Because the script uses
the absolute path of the executable, it is enough to use instdir/program
as the current working directory.

The symptom was that in the above case it was not possible to run the
workbenches like vcldemo, minvcl, etc. The error was:

$ ./bin/run minvcl

minvcl.exe - Entry Point Not Found | The procedure entry point
?RasterFromBitmap@SKImages@@... could not be located in the dynamic
link library vcllo.dll

$ ldd ./lode/dev/core/workdir/LinkTarget/Executable/minvcl.exe
	...
        comphelper.dll => not found
        fwklo.dll => not found
        i18nlangtag.dll => /cygdrive/c/Program Files/LibreOffice/program/i18nlangtag.dll (?)
        sal3.dll => /cygdrive/c/Program Files/LibreOffice/program/sal3.dll (?)
        vcllo.dll => not found
        MSVCP140D.dll => /cygdrive/c/WINDOWS/SYSTEM32/MSVCP140D.dll (?)
        VCRUNTIME140D.dll => /cygdrive/c/WINDOWS/SYSTEM32/VCRUNTIME140D.dll (?)
        VCRUNTIME140_1D.dll => /cygdrive/c/WINDOWS/SYSTEM32/VCRUNTIME140_1D.dll (?)
        ucrtbased.dll => /cygdrive/c/WINDOWS/SYSTEM32/ucrtbased.dll (?)

Change-Id: I02b6d9aca1062b22a91d5a1f9bce0cf7238a896a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156902
Tested-by: Jenkins
Reviewed-by: Hossein <hossein@libreoffice.org>