Related: tdf#55160 sw floattable, nested DOCX exp: fix inner tbl at cell start
A cut-down bugdoc had an inline table, and an inner floating table,
anchored in the start of the C1 cell.

Exporting to DOCX resulted in a layout that looks like the floating
table is not anchored inside the outer table anymore. Checking the
markup, the floating table was written between row 1 & row 2, which is
not a valid position for a floating table. Probably the intention was to
write the floating table before the first paragraph in C1.

Fix the problem by still writing the floating table in
DocxAttributeOutput::StartParagraph(), before opening <w:p>, but do this
after opening table/row/cell, which is late enough to have a correct
anchor but is early enough to be still outside the paragraph.

The import side of this still needs fixing.

Change-Id: I1827b2def1faaf53312522f5a8415fbe697fcb8b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156904
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins