tdf#71224: Remove "Insert fields" option for an existing table (workaround)
For the moment, we don't implement insert field at specific position
It's not SQL standard and each database has made its choice (some use "BEFORE", other "FIRST" and "AFTER")
and some, like Postgresql, don't allow this.
So for the moment, test if the table already exists (and so it's an edition), in this case only
we remove "Insert fields" entry. Indeed, in case of new table, there's no pb.

The real fix is to implement the insert for each database + error message for those which don't support this

Change-Id: I27010ed2dddee3f8746245406f2a0cb629b3d8e6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156948
Reviewed-by: Julien Nabet <serval2412@yahoo.fr>
Tested-by: Jenkins